unhinged performance improvement find+replace
I set a precedent for using entities for variable storage in 2019. We're in pretty deep, so I'm going to just try and optimize what we have instead of doing 9 months of work all over again.

Selectors that involve entity variable holders are now chunkLimited to improve query performance. This makes the datapack considerably chunkier, but the runtime performance should be better.